feat(KVNamespace): bring kv implementation more in line with docs
This commit adds value types to get and put, the getWithMetadata and list functions, and returns null if a key could not be found. This brings the KV implementation more in line with the documentation here: https://developers.cloudflare.com/workers/runtime-apis/kv. The only things missing now are ReadableStream & FormData support, and potentially errors when data is too large.