feat(FileKVStore): enable file system backed kv store by env
This allows the file system backed KV store to be used for all KV namespaces, not just for Workers Sites. Setting the "KV_FILE_ROOT" environment variable enables this. For namespace "NAMESPACE", files will be stored in "KV_FILE_ROOT/NAMESPACE". Values will be stored in files named the same as their key. Metadata and expiration information (if provided) is stored in a .meta.json file prefixed with the key.